test: 검증 로직 테스트를 ValidatorTest로 이동하여 구조 개선
